Displays random quote each time the app launches
