--- OneYes/OneYes/Resources/ModelData.swift
@@ -1,33 +1,0 @@
-//
-//  ModelData.swift
-//  OneYes
-//
-//  Created by iMac Pro on 4/3/23.
-//
-
-import Foundation
-
-class ModelData {
-    var quotesData: [Quote] = load("quotesData.json")
-}
-
-func load<T: Decodable>(_ filename: String) -> T {
-    let data: Data
-
-    guard let file = Bundle.main.url(forResource: filename, withExtension: nil) else {
-        fatalError("Couldn't find \(filename) in main bundle.")
-    }
-
-    do {
-        data = try Data(contentsOf: file)
-    } catch {
-        fatalError("Couldn't load \(filename) from main bundle:\n\(error)")
-    }
-
-    do {
-        let decoder = JSONDecoder()
-        return try decoder.decode(T.self, from: data)
-    } catch {
-        fatalError("Couldn't parse \(filename) as \(T.self):\n\(error)")
-    }
-}
