+++ OneYes/OneYes/Storyboards/Base.lproj/Main.storyboard
@@ -2,6 +2,7 @@
 <document type="com.apple.InterfaceBuilder3.CocoaTouch.Storyboard.XIB" version="3.0" toolsVersion="21701" targetRuntime="iOS.CocoaTouch" propertyAccessControl="none" useAutolayout="YES" useTraitCollections="YES" useSafeAreas="YES" colorMatched="YES" initialViewController="RiO-nV-obI">
     <device id="retina4_7" orientation="portrait" appearance="light"/>
     <dependencies>
+        <deployment identifier="iOS"/>
         <plugIn identifier="com.apple.InterfaceBuilder.IBCocoaTouchPlugin" version="21678"/>
         <capability name="Safe area layout guides" minToolsVersion="9.0"/>
         <capability name="Stack View standard spacing" minToolsVersion="9.0"/>
@@ -180,7 +181,7 @@
             <objects>
                 <viewController id="VKn-yb-ZrU" customClass="NoDetailViewController" customModule="OneYes" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="SEJ-Il-jFk">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="647"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <stackView opaque="NO" contentMode="scaleToFill" spacingType="standard" translatesAutoresizingMaskIntoConstraints="NO" id="fSX-Yj-W8o">
@@ -366,17 +367,17 @@
                         <rect key="frame" x="0.0" y="0.0" width="359" height="130"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
-                            <label opaque="NO" userInteractionEnabled="NO" contentMode="left" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Quote of the Day Placeholder" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Ofv-ru-wbY">
-                                <rect key="frame" x="12" y="20" width="335" height="57"/>
-                                <fontDescription key="fontDescription" type="system" pointSize="17"/>
+                            <label opaque="NO" userInteractionEnabled="NO" contentMode="TopLeft" horizontalHuggingPriority="251" verticalHuggingPriority="251" text="Quote of the Day Placeholder" textAlignment="natural" lineBreakMode="tailTruncation" numberOfLines="0" baselineAdjustment="alignBaselines" adjustsFontSizeToFit="NO" translatesAutoresizingMaskIntoConstraints="NO" id="Ofv-ru-wbY">
+                                <rect key="frame" x="12" y="2" width="335" height="76"/>
+                                <fontDescription key="fontDescription" type="system" pointSize="15"/>
                                 <nil key="textColor"/>
                                 <nil key="highlightedColor"/>
                             </label>
                             <textField opaque="NO" contentMode="scaleToFill" horizontalHuggingPriority="248" contentHorizontalAlignment="left" contentVerticalAlignment="center" borderStyle="roundedRect" placeholder="Today, I need a yes for..." textAlignment="natural" minimumFontSize="17" clearButtonMode="whileEditing" translatesAutoresizingMaskIntoConstraints="NO" id="xNR-3p-ebz">
-                                <rect key="frame" x="12" y="77" width="335" height="45"/>
+                                <rect key="frame" x="12" y="80" width="335" height="42"/>
                                 <color key="backgroundColor" systemColor="systemGray6Color"/>
                                 <constraints>
-                                    <constraint firstAttribute="height" constant="45" id="B33-IE-TJv"/>
+                                    <constraint firstAttribute="height" constant="42" id="B33-IE-TJv"/>
                                 </constraints>
                                 <fontDescription key="fontDescription" type="system" pointSize="14"/>
                                 <textInputTraits key="textInputTraits"/>
@@ -387,14 +388,15 @@
                         <constraints>
                             <constraint firstItem="DgT-wH-0OZ" firstAttribute="bottom" secondItem="xNR-3p-ebz" secondAttribute="bottom" constant="8" id="5VJ-VP-DJl"/>
                             <constraint firstItem="DgT-wH-0OZ" firstAttribute="trailing" secondItem="Ofv-ru-wbY" secondAttribute="trailing" constant="12" id="8Hr-Pc-emR"/>
-                            <constraint firstItem="xNR-3p-ebz" firstAttribute="top" secondItem="Ofv-ru-wbY" secondAttribute="bottom" id="Oq5-At-zmA"/>
+                            <constraint firstItem="xNR-3p-ebz" firstAttribute="top" secondItem="Ofv-ru-wbY" secondAttribute="bottom" constant="2" id="Oq5-At-zmA"/>
                             <constraint firstItem="Ofv-ru-wbY" firstAttribute="leading" secondItem="DgT-wH-0OZ" secondAttribute="leading" constant="12" id="duu-2t-6AZ"/>
                             <constraint firstItem="xNR-3p-ebz" firstAttribute="leading" secondItem="DgT-wH-0OZ" secondAttribute="leading" constant="12" id="rHq-6B-1Df"/>
-                            <constraint firstItem="Ofv-ru-wbY" firstAttribute="top" secondItem="DgT-wH-0OZ" secondAttribute="top" id="xhc-eB-rf8"/>
+                            <constraint firstItem="Ofv-ru-wbY" firstAttribute="top" secondItem="4Bv-MC-v2o" secondAttribute="top" constant="2" id="xhc-eB-rf8"/>
                             <constraint firstItem="DgT-wH-0OZ" firstAttribute="trailing" secondItem="xNR-3p-ebz" secondAttribute="trailing" constant="12" id="xjb-aB-fBi"/>
                         </constraints>
                     </view>
                     <connections>
+                        <outlet property="dailyQuoteLabel" destination="Ofv-ru-wbY" id="T7P-Rl-7j1"/>
                         <outlet property="needYesTextField" destination="xNR-3p-ebz" id="EDf-nf-K7T"/>
                     </connections>
                 </viewController>
@@ -532,7 +534,7 @@
             <objects>
                 <viewController id="rLr-1u-r0a" customClass="YesDetailViewController" customModule="OneYes" customModuleProvider="target" sceneMemberID="viewController">
                     <view key="view" contentMode="scaleToFill" id="JNv-fP-2TZ">
-                        <rect key="frame" x="0.0" y="0.0" width="375" height="667"/>
+                        <rect key="frame" x="0.0" y="0.0" width="375" height="647"/>
                         <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
                         <subviews>
                             <stackView opaque="NO" contentMode="scaleToFill" spacingType="standard" translatesAutoresizingMaskIntoConstraints="NO" id="1ZW-8Y-woX">
