+++ OneYes/OneYes/UI/Home View/Quote and Need/QuoteAndNeedYesViewModel.swift
@@ -5,13 +5,14 @@
 //  Created by iMac Pro on 4/5/23.
 //
 
-import Foundation
+import UIKit
 
 struct QuoteAndNeedYesViewModel {
     
     //MARK: - PROPERTIES
     var reason: Reason?
     let service: FirebaseService
+    var quotesData: [Quote] = QuoteDataService.load("quotesData.json")
         
     init(reason: Reason? = nil, service: FirebaseService = FirebaseService()) {
         self.reason   = reason
@@ -23,5 +24,12 @@ struct QuoteAndNeedYesViewModel {
         service.saveNewReasonToFirestore(title: reasonTitle, logs: logs) {
             NotificationCenter.default.post(name: Constants.Notifications.newReasonCreated, object: nil)
         }
+    }
+    
+    func loadRandomQuote(with quoteString: UILabel) {
+        var index = Int.random(in: 0..<quotesData.count)
+        let quote = quotesData[index].quote
+        let author = quotesData[index].author
+        quoteString.text = "\"\(quote)\" -\(author)"
     }
 }
